fix yarn start error
